Trying the break the code into smaller chunks
Separated RetreieveData and made is its own class.
Created class for accessing the directory with the argument fileName.
Formed a class for Disney
Uploaded separate directories for all offered streaming services.
